requirements.txt
# Amazon Price Monitor Dependencies

# Database
mysql-connector-python>=8.0.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Environment variables
python-dotenv>=1.0.0

# CLI formatting
tabulate>=0.9.0

# Optional: For scheduled tasks
# schedule>=1.2.0

# Optional: For async operations
# aiohttp>=3.8.0
